Adds follow button to open login modal if user is not logged in
