+++ app/components/user/UserFollowButton.tsx
@@ -1,16 +1,34 @@
 "use client";
 
+import axios from "axios";
+
+import { useCallback } from "react";
+import { toast } from "react-toastify";
+
 import { SafeUser } from "@/app/types";
 
 interface UserFollowButtonProps {
-  userFollowList: SafeUser[];
+  currentUser: SafeUser;
   userId: string;
+  username: string;
 }
 
 const UserFollowButton: React.FC<UserFollowButtonProps> = ({
-  userFollowList,
+  currentUser,
   userId,
+  username,
 }) => {
+  const onFollow = useCallback(() => {
+    axios
+      .post("/api/follow", { userId })
+      .then(() => {
+        toast(`Followed ${username}`);
+      })
+      .catch((error) => {
+        toast(error);
+      });
+  }, [userId, currentUser.following]);
+
   return <div></div>;
 };
